Proof-of-concept
I don't fully understand viewBox and sizing.  It's also not clear what the
tile size should be -- should we instead use percentages of the full
image (assuming tiles are equal size), for responsive sizing?  What if the
viewport isn't square (or the tiles themselves are not square)?  We're
also dependent on external files providing sizing -- the implicit size of
100% probably does not work here.

Batik isn't so great -- there should be a way to get actual SVG nodes with
setters and stuff, not the generic DOM interface.

It'd be nice for mappings to apply styles to external files, for creating
e.g. red and blue players in games out of a single (black) set of pieces.

We should add an --output <file> option that writes SVG, PDF or some
raster format (PNG).